usermanagement/requirements.txt
asgiref
autobahn
bcrypt
psycopg2-binary
certifi
channels
chardet
click
colorama
cryptography
daphne
django
djangorestframework
docker
docker-compose
docopt
fasteners
filelock
future
httplib2
idna
importlib-metadata
incremental
jsonschema
keyring
monotonic
more-itertools
mysql-connector
mysql-connector-python
netifaces
oauthlib
paramiko
pika
Pillow
protobuf
psutil
PyJWT
pymacaroons
PyNaCl
pyOpenSSL
pycparser
PyYAML
requests
requests-toolbelt
SecretStorage
six
sqlparse
texttable
typing_extensions
urllib3
websocket-client
zipp
